backend/requirements.txt
Flask
flask-cors
numpy
pillow
opencv-python
matplotlib
